+++ native-windows-gui/src/win32/image_decoder.rs
@@ -73,7 +73,6 @@ pub unsafe fn create_decoder_from_stream(fact: &IWICImagingFactory, data: &mut [
     );
 
     if r != S_OK {
-        println!("{:X}", r);
         return Err(NwgError::resource_create("Failed to create decoder from stream"));
     }